requirements.txt
fastapi[standard]>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
ag2>=0.3.2
openai>=1.40.0
python-dotenv
google-ads